99% done algorithm
